added constructor for strings

+++ GreatBin.hpp
@@ -12,7 +12,7 @@ struct GreatBin {
     // helper function for the division algorithm
     GreatBin find_beta(const GreatBin&,const GreatBin&, GreatBin*) const;
     // modulus of a string
-    int str_mod(std::string&, long);
+    long str_mod(std::string&, long);
     // division of a string
     std::string str_div(std::string&, long);
     bool str_greater (std::string&, std::string&);
